+++ pages/_app.tsx
@@ -21,7 +21,9 @@ import { useRouter } from "next/router";
 import { GetServerSideProps } from "next";
 import aaUtils from "../src/externals/aaUtils";
 import Script from "next/script";
+import { theme } from "../src/theme";
 import pkg from "../package.json";
+
 interface MyAppProps {
   Component: any;
   pageProps: any;
@@ -173,7 +175,7 @@ function MyApp({ Component, pageProps }: MyAppProps) {
           `,
         }}
       />
-      <DSProvider>
+      <DSProvider theme={theme}>
         <FormProvider {...formMethods}>
           <FormDataContextProvider initState={initState}>
             <ApplicationContainer problemDetail={error}>